requirements.txt
Django==3.1.1
django-ckeditor==6.0.0
Pillow==7.2.0
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
